+++ examples/with-vite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # with-vite
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [a1d6776]
+  - @rainbow-me/rainbowkit@0.8.1
+
 ## 0.0.19
 
 ### Patch Changes