+++ packages/create-rainbowkit/templates/next-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rainbowkit-next-app
 
+## 0.0.34
+
+### Patch Changes
+
+- Updated dependencies [a1d6776]
+  - @rainbow-me/rainbowkit@0.8.1
+
 ## 0.0.33
 
 ### Patch Changes